Manual

Manual

Defining tests

The entry-point file (commonly called runtests.jl) is simply:

using Jute
exit(runtests())

The test runner picks up any file with the name ending in .test.jl in the directory where the entry-point file is located, or in any subdirectories. All those files are included at the same level (with using Jute at the start), and all the @testcase and @testgroup definitions are picked up. The @testgroup definitions can contain other @testgroup definitions and @testcase definitions.

The exit() call is required to signal about any test failures to the processes that initiate the execution of the test suite, for instance CI tools. runtests() returns 1 if there were failed tests, 0 otherwise.

The @testcase macro takes the testcase name and body:

@testcase "simple testcase" begin
    @test 1 == 1
end

Assertions

Jute relies on the assertions from Base.Test; @test, @test_throws, @test_skip and @test_broken can be used. In addition, Jute has a @test_result macro allowing one to return a custom result (e.g. the value of a benchmark from a testcase). There can be several assertions per testcase; their results will be reported separately. If the testcase does not call any assertions and does not throw any exceptions, it is considered to be passed.

Grouping tests

Testcases can be grouped using @testgroup definitions. For example:

@testcase "tc1" begin
end

@testgroup "group" begin
    @testcase "tc2" begin
    end
end

@testgroup "group2" begin
    @testgroup "subgroup" begin
        @testcase "tc3" begin
        end
    end
end

the following testcases will be listed:

tc1
Group/tc2
Group2/Subgroup/tc3

The order of testcase definition is preserved. In other words, the testcases will be executed in the same order in which they were defined.

Parametrizing testcases

Constant fixtures

The simplest method to parametrize a test is to supply it with an iterable:

@testcase "parametrized testcase" for x in [1, 2, 3]
    @test x == 1
end

# Output:
# parametrized testcase[1]: [PASS]
# parametrized testcase[2]: [FAIL]
# parametrized testcase[3]: [FAIL]

By default, Jute uses string() to convert a fixture value to a string for reporting purposes. One can assign custom labels to fixtures by passing a Pair of iterables instead:

@testcase "parametrized testcase" for x in [1, 2, 3] => ["one", "two", "three"]
    @test x == 1
end

# Output:
# parametrized testcase[one]: [PASS]
# parametrized testcase[two]: [FAIL]
# parametrized testcase[three]: [FAIL]

A testcase can use several fixtures, in which case Jute will run the testcase function will all possible combinations of them:

@testcase "parametrized testcase" for x in [1, 2], y in [3, 4]
    @test x + y == y + x
end

# Output:
# parametrized testcase[1, 3]: [PASS]
# parametrized testcase[1, 4]: [PASS]
# parametrized testcase[2, 3]: [PASS]
# parametrized testcase[2, 4]: [PASS]

Global fixtures

A global fixture is a more sophisticated variant of a constant fixture that has a setup and a teardown stage. For each global fixture, the setup is called before the first testcase that uses it. As for the teardown, it is either called right away (if the keyword parameter instant_teardown is true), or after the last testcase that uses it (if instant_teardown is false, which is the default). If no testcases use it (for example, they were filtered out), neither setup nor teardown will be called.

The setup and the teardown are defined by use of a single coroutine that produces the fixture iterable. The coroutine's first argument is a function that is used to return the fixture values. If instant_teardown is false, the call blocks until it is time to execute the teardown:

db_connection = fixture() do produce
    c = db_connect()

    # this call blocks until all the testcases
    # that use the fixture are executed
    produce([c])

    close(c)
end

Note that a global fixture must produce the whole iterable in one go.

Similarly to the constant fixture case, one can provide a custom identifier for the fixture via the optional second argument of produce():

db_connection = fixture() do produce
    c = db_connect()

    # this call blocks until all the testcases
    # that use the fixture are executed
    produce([c], ["db_connection"])

    close(c)
end

Global fixtures can be parametrized by other constant or global fixtures. Similarly to the test parametrization, all possible combinations of parameters will be used to produce iterables, which will be chained together:

fx1 = fixture() do produce
    produce(3:4)
end

fx2 = fixture(1:2, fx1) do produce, x, y
    produce([(x, y)])
end

@testcase "tc" for x in fx2
    @test length(x) == 2
end

# Output:
# tc[(1, 3)]: [PASS]
# tc[(1, 4)]: [PASS]
# tc[(2, 3)]: [PASS]
# tc[(2, 4)]: [PASS]

Local fixtures

A local fixture is a fixture whose value is created right before each call to the testcase function and destroyed afterwards. A simple example is a fixture that provides a temporary directory:

temporary_dir = local_fixture() do produce
    dir = mktempdir()
    produce(dir) # this call will block while the testcase is being executed
    rm(dir, recursive=true)
end

@testcase "tempdir test" for dir in temporary_dir
    open(joinpath(dir, "somefile"), "w")
end

Note that, unlike a global fixture, a local fixture only produces one value. Local fixtures can be parametrized by any other type of fixture, including other local fixtures.

Testcase tags

Warning

Tagging implementation is a work in progress for testcase macros.

Testcases can be assigned tags of the type Symbol. This can be used to establish a secondary grouping, independent of the primary grouping provided by modules. For example, one can tag performance tests, tests that run for a long time, unit/integration tests, tests that require a specific resource and so on. Testcases can be filtered by tags they have or don't have using command-line arguments.

The tagging is performed by the function tag() that takes a Symbol and returns a function that tags a testcase:

tc = tag(:foo)(testcase() do
    ... something
end)

It is convenient to use the <| operator:

tc =
    tag(:foo) <|
    testcase() do
        ... something
    end

A tag can be removed from a testcase using untag. Note that tagging and untagging commands are applied from inner to outer, so, for example, the following code

tc =
    tag(:foo) <|
    untag(:bar) <|
    untag(:foo) <|
    tag(:bar) <|
    testcase() do
        ... something
    end

will leave tc with the tag :foo, but without the tag :bar.

Run options

Jute's runtest() picks up the options from the command line by default. Alternatively, they can be set with the options keyword argument of runtests().

Jute.build_parser — Function.

For every option, the corresponding command-line argument names are given in parentheses. If supplied via the options keyword argument of runtests(), their type must be as given or convert()-able to it.

:include_only:: Nullable{Regex} (--include-only, -i): takes a regular expression; tests with full names that do not match it will not be executed.

:exclude:: Nullable{Regex} (--exclude, -e): takes a regular expression; tests with full names that match it will not be executed.

:verbosuty:: Int (--verbosity, -v): 0, 1 or 2, defines the amount of output that will be shown. 1 is the default.

:include_only_tags:: Array{Symbol, 1} (--include-only-tags, -t): include only tests with any of the specified tags. You can pass several tags to this option, separated by spaces.

:exclude_tags:: Array{Symbol, 1} (--exclude-tags, -t): exclude tests with any of the specified tags. You can pass several tags to this option, separated by spaces.

:max_fails:: Int (--max-fails): stop after the given amount of failed testcases (a testcase is considered failed, if at least one test in it failed, or an unhandeld exception was thrown).

:capture_output:: Bool (--capture-output): capture all the output from testcases and only show the output of the failed ones in the end of the test run.

:dont_add_runtests_path::: Bool (`–dont-add-runtests-path): capture testcase output and display only the output from failed testcases after all the testcases are finished.

:test_file_postifx:: String (--test-file-postfix): postfix of the files which will be picked up by the automatic testcase discovery.

source

Run options can be accessed from a testcase or a fixture via the built-in fixture run_options.