Public API

Public API

Entry point

Jute.runtests — Function.

runtests(; options=nothing)

Run the test suite.

This function has several side effects:

it parses the command-line arguments, using them to build the dictionary of run options (see Run options in the manual for the list);

it picks up and includes the test files, selected according to the options.

options must be a dictionary with the keys corresponding to some of the options from the above list. If options is given, command-line arguments are not parsed.

Returns 0 if there are no failed tests, 1 otherwise.

source

Testcases and fixtures

Jute.testcase — Function.

testcase(func, params...)

Define a testcase.

func is a testcase function. The number of function parameters must be equal to the number of parametrizing fixtures given in params. This function will be called with all combinations of values of fixtures from params.

params are either fixtures, iterables or pairs of two iterables used to parametrize the function. In the latter case, the first iterable will be used to produce the values, and the second one to produce the corresponding labels (for logging).

Returns a Testcase object.

source

Jute.fixture — Function.

fixture(func, params...; instant_teardown=false)

Create a global fixture (a fixture set up once before all the testcases that use it and torn down after they finish).

func is a function with length(params) + 1 parameters. The first parameter takes a function produce(values[, labels]) that is used to return the fixture iterable (with an optional iterable of labels). The rest take the values of the dependent fixtures from params.

params are either fixtures (constant of global only), iterables or pairs of two iterables used to parametrize the fixture.

Returns a GlobalFixture object.

source

Jute.local_fixture — Function.

local_fixture(func, params...)

Create a local fixture (a fixture set up before each testcase that uses it and torn down afterwards).

func is a function with length(params) + 1 parameters. The first parameter takes a function produce(value[, label]) that is used to return the fixture value (with an optional label). The rest take the values of the dependent fixtures from params.

params are either fixtures (of any type), iterables or pairs of two iterables used to parametrize the fixture.

Returns a LocalFixture object.

source

Assertions

The following assertions are re-exported from Base.Test and can be used inside Jute testcases.

Base.Test.@test — Macro.

@test ex
@test f(args...) key=val ...

Tests that the expression ex evaluates to true. Returns a Pass Result if it does, a Fail Result if it is false, and an Error Result if it could not be evaluated.

The @test f(args...) key=val... form is equivalent to writing @test f(args..., key=val...) which can be useful when the expression is a call using infix syntax such as approximate comparisons:

@test a ≈ b atol=ε

This is equivalent to the uglier test @test ≈(a, b, atol=ε). It is an error to supply more than one expression unless the first is a call expression and the rest are assignments (k=v).

source

Base.Test.@test_throws — Macro.

@test_throws exception expr

Tests that the expression expr throws exception. The exception may specify either a type, or a value (which will be tested for equality by comparing fields). Note that @test_throws does not support a trailing keyword form.

source

Base.Test.@test_broken — Macro.

@test_broken ex
@test_broken f(args...) key=val ...

Indicates a test that should pass but currently consistently fails. Tests that the expression ex evaluates to false or causes an exception. Returns a Broken Result if it does, or an Error Result if the expression evaluates to true.

The @test_broken f(args...) key=val... form works as for the @test macro.

source

Base.Test.@test_skip — Macro.

@test_skip ex
@test_skip f(args...) key=val ...

Marks a test that should not be executed but should be included in test summary reporting as Broken. This can be useful for tests that intermittently fail, or tests of not-yet-implemented functionality.

The @test_skip f(args...) key=val... form works as for the @test macro.

source

This is an additional assertion, allowing one to record an arbitrary value as a test result.

Jute.@test_result — Macro.

@test_result expr

Records a result from the test. The result of expr will be displayed in the report by calling string() on it.

source

Testcase tags

Jute.tag — Function.

tag(::Symbol)

Returns a function that tags a testcase with the given tag:

tc = tag(:foo)(testcase() do
    ... something
end)

Testcases can be filtered in/out using run options. It is convenient to use the <| operator:

tc =
    tag(:foo) <|
    testcase() do
        ... something
    end

Note that tag and untag commands are applied from inner to outer.

source

Jute.untag — Function.

untag(::Symbol)

Returns a function that untags a testcase with the given tag. See tag for more details.

source

Jute.:<| — Function.

<|(f, x) === f(x)

A helper operator that makes applying testcase tags slightly more graceful. See tag for an example.

source

Built-in fixtures

Jute.temporary_dir — Constant.

A local fixture that creates a temporary directory and returns its name; the directory and all its contents is removed during the teardown.

source

Jute.run_options — Constant.

A global fixture that returns the dictionary with the current run options (see Run options in the manual for the full list.

source

Utilities

Jute.rowmajor_product — Function.

rowmajor_product(xss...)

Iterate over all combinations in the cartesian product of the inputs. Similar to IterTools.product(), but iterates in row-major order (that is, the first iterator is iterated the slowest).

source

Jute.pprint_time — Function.

pprint_time(s; meaningful_digits=0)

Returns a string that represents a given time (in seconds) as a value scaled to the appropriate unit (minutes, hours, milliseconds etc) and rounded to a given number of meaningful digits (if it is smaller than a minute). If the latter is 0, the result is rounded to an integer at all times.

source

Jute.with_output_capture — Function.

with_output_capture(func, pass_through=false)

Execute the callable func and capture its output (both STDOUT and STDERR) in a string. Returns a tuple of the func's return value and its output. If pass_through is true, does not capture anything and returns an empty string instead of the output.

source