Public API

Public API

Entry point

Jute.runtests — Function.

runtests(; options=nothing)

Run the test suite.

This function has several side effects:

it parses the command-line arguments, using them to build the dictionary of run options (see Run options in the manual for the list);

it picks up and includes the test files, selected according to the options.

options must be a dictionary with the keys corresponding to some of the options from the above list. If options is given, command-line arguments are not parsed.

Returns 0 if there are no failed tests, 1 otherwise.

source

Testcases and fixtures

Jute.@testcase — Macro.

@testcase [option=val ...] <name> begin ... end
@testcase [option=val ...] <name> for x in fx1, y in fx2 ... end

Create a testcase object and add it to the current test group.

Available options:

tags :: Array{Symbol, 1}: a list of tags for the testcase.

source

Jute.@testgroup — Macro.

@testgroup <name> begin ... end

Create a test group. The body can contain other @testgroup or @testcase declarations.

source

Jute.fixture — Function.

fixture(func, params...; instant_teardown=false)

Create a global fixture (a fixture set up once before all the testcases that use it and torn down after they finish).

func is a function with length(params) + 1 parameters. The first parameter takes a function produce(values[, labels]) that is used to return the fixture iterable (with an optional iterable of labels). The rest take the values of the dependent fixtures from params.

params are either fixtures (constant of global only), iterables or pairs of two iterables used to parametrize the fixture.

Returns a GlobalFixture object.

source

Jute.local_fixture — Function.

local_fixture(func, params...)

Create a local fixture (a fixture set up before each testcase that uses it and torn down afterwards).

func is a function with length(params) + 1 parameters. The first parameter takes a function produce(value[, label]) that is used to return the fixture value (with an optional label). The rest take the values of the dependent fixtures from params.

params are either fixtures (of any type), iterables or pairs of two iterables used to parametrize the fixture.

Returns a LocalFixture object.

source

Assertions

The following assertions are re-exported from Base.Test and can be used inside Jute testcases.

Base.Test.@test — Macro.

@test ex
@test f(args...) key=val ...

Tests that the expression ex evaluates to true. Returns a Pass Result if it does, a Fail Result if it is false, and an Error Result if it could not be evaluated.

The @test f(args...) key=val... form is equivalent to writing @test f(args..., key=val...) which can be useful when the expression is a call using infix syntax such as approximate comparisons:

@test a ≈ b atol=ε

This is equivalent to the uglier test @test ≈(a, b, atol=ε). It is an error to supply more than one expression unless the first is a call expression and the rest are assignments (k=v).

source

Base.Test.@test_throws — Macro.

@test_throws exception expr

Tests that the expression expr throws exception. The exception may specify either a type, or a value (which will be tested for equality by comparing fields). Note that @test_throws does not support a trailing keyword form.

source

Base.Test.@test_broken — Macro.

@test_broken ex
@test_broken f(args...) key=val ...

Indicates a test that should pass but currently consistently fails. Tests that the expression ex evaluates to false or causes an exception. Returns a Broken Result if it does, or an Error Result if the expression evaluates to true.

The @test_broken f(args...) key=val... form works as for the @test macro.

source

Base.Test.@test_skip — Macro.

@test_skip ex
@test_skip f(args...) key=val ...

Marks a test that should not be executed but should be included in test summary reporting as Broken. This can be useful for tests that intermittently fail, or tests of not-yet-implemented functionality.

The @test_skip f(args...) key=val... form works as for the @test macro.

source

Base.Test.@inferred — Macro.

@inferred f(x)

Tests that the call expression f(x) returns a value of the same type inferred by the compiler. It is useful to check for type stability.

f(x) can be any call expression. Returns the result of f(x) if the types match, and an Error Result if it finds different types.

julia> using Base.Test

julia> f(a,b,c) = b > 1 ? 1 : 1.0
f (generic function with 1 method)

julia> typeof(f(1,2,3))
Int64

julia> @code_warntype f(1,2,3)
Variables:
  #self#::#f
  a::Int64
  b::Int64
  c::Int64

Body:
  begin
      unless (Base.slt_int)(1, b::Int64)::Bool goto 3
      return 1
      3:
      return 1.0
  end::UNION{FLOAT64, INT64}

julia> @inferred f(1,2,3)
ERROR: return type Int64 does not match inferred return type Union{Float64, Int64}
Stacktrace:
 [1] error(::String) at ./error.jl:33

julia> @inferred max(1,2)
2

source

Base.Test.@test_warn — Macro.

@test_warn msg expr

Test whether evaluating expr results in STDERR output that contains the msg string or matches the msg regular expression. If msg is a boolean function, tests whether msg(output) returns true. If msg is a tuple or array, checks that the error output contains/matches each item in msg. Returns the result of evaluating expr.

See also @test_nowarn to check for the absence of error output.

source

Base.Test.@test_nowarn — Macro.

@test_nowarn expr

Test whether evaluating expr results in empty STDERR output (no warnings or other messages). Returns the result of evaluating expr.

source

Jute adds several assertions of its own.

Jute.@test_result — Macro.

@test_result expr

Records a result from the test. The result of expr will be displayed in the report by calling string() on it.

source

Jute.@test_fail — Macro.

@test_fail descr

Report a fail, providing an additional description (must be convertable to String). The description will be displayed in the final report at the end of the test run.

source

Built-in fixtures

Jute.temporary_dir — Constant.

A local fixture that creates a temporary directory and returns its name; the directory and all its contents is removed during the teardown.

source

Jute.run_options — Constant.

A global fixture that returns the dictionary with the current run options (see Run options in the manual for the full list.

source